AnitaRasa-main/requirements.txt
wheel==0.45.1
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.3.1
anyio==4.5.2
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.17.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
blis==0.7.11
boto3==1.36.16
botocore==1.36.16
breadability==0.1.20
Brotli==1.1.0
CacheControl==0.12.14
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.5
confluent-kafka==2.8.0
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.11
cytoolz==1.0.1
dask==2022.10.2
dateparser==1.2.0
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.2.2
fbmessenger==6.0.0
filelock==3.16.1
fire==0.7.0
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gast==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-cloud-vision==3.10.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
jellyfish==1.1.3
Jinja2==3.1.5
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.3.0
libclang==18.1.1
locket==1.0.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
msgpack==1.1.0
multidict==5.2.0
murmurhash==1.0.12
networkx==2.6.3
nltk==3.9.1
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pathlib_abc==0.1.1
pathy==0.11.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
prompt-toolkit==3.0.28
propcache==0.2.0
proto-plus==1.26.1
protobuf==4.23.3
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==1.10.9
pydot==1.4.2
pydyf==0.11.0
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.4
pypdfium2==4.30.1
pyphen==0.16.0
pyreadline3==3.5.4
pyrsistent==0.20.0
pyspellchecker==0.8.2
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2022.7
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.11.2
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
sib-api-v3-sdk==7.6.0
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.34.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
srsly==2.4.8
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.1.12
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.3
toolz==1.0.0
tqdm==4.67.1
twilio==8.2.2
typer==0.7.0
typing-utils==0.1.0
typing_extensions==4.5.0
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
wasabi==1.1.3
wcwidth==0.2.13
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.6
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.10
yarl==1.15.2
zipp==3.20.2
zopfli==0.2.3.post1
